requirements.txt
jsonschema==4.19.1
openai==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
PyYAML==6.0.1
redis==5.0.1
structlog==23.2.0
tenacity==8.2.3
tzlocal==5.2